Improve yoo
