+++ cmake-build-debug/.cmake/api/v1/reply/target-yo_dwm-Debug-631731075c5f906219d9.json
@@ -11,6 +11,7 @@
 		"commands" : 
 		[
 			"add_executable",
+			"target_link_libraries",
 			"include_directories"
 		],
 		"files" : 
@@ -31,23 +32,29 @@
 			{
 				"command" : 1,
 				"file" : 0,
+				"line" : 31,
+				"parent" : 0
+			},
+			{
+				"command" : 2,
+				"file" : 0,
 				"line" : 6,
 				"parent" : 0
 			},
 			{
-				"command" : 1,
+				"command" : 2,
 				"file" : 0,
 				"line" : 7,
 				"parent" : 0
 			},
 			{
-				"command" : 1,
+				"command" : 2,
 				"file" : 0,
 				"line" : 8,
 				"parent" : 0
 			},
 			{
-				"command" : 1,
+				"command" : 2,
 				"file" : 0,
 				"line" : 9,
 				"parent" : 0
@@ -69,19 +76,19 @@
 			"includes" : 
 			[
 				{
-					"backtrace" : 2,
+					"backtrace" : 3,
 					"path" : "/home/anas/suckless_utils/yo-dwm/."
 				},
 				{
-					"backtrace" : 3,
+					"backtrace" : 4,
 					"path" : "/home/anas/suckless_utils/yo-dwm/src"
 				},
 				{
-					"backtrace" : 4,
+					"backtrace" : 5,
 					"path" : "/home/anas/suckless_utils/yo-dwm/src/keys"
 				},
 				{
-					"backtrace" : 5,
+					"backtrace" : 6,
 					"path" : "/home/anas/suckless_utils/yo-dwm/src/layouts"
 				}
 			],
@@ -106,7 +113,8 @@
 				13,
 				14,
 				15,
-				16
+				16,
+				18
 			]
 		}
 	],
@@ -122,6 +130,21 @@
 			{
 				"fragment" : "",
 				"role" : "flags"
+			},
+			{
+				"backtrace" : 2,
+				"fragment" : "-lX11",
+				"role" : "libraries"
+			},
+			{
+				"backtrace" : 2,
+				"fragment" : "-lXft",
+				"role" : "libraries"
+			},
+			{
+				"backtrace" : 2,
+				"fragment" : "-lXinerama",
+				"role" : "libraries"
 			}
 		],
 		"language" : "C"
@@ -149,7 +172,8 @@
 				13,
 				14,
 				15,
-				16
+				16,
+				18
 			]
 		},
 		{
@@ -162,7 +186,10 @@
 				9,
 				10,
 				12,
-				17
+				17,
+				19,
+				20,
+				21
 			]
 		}
 	],
@@ -267,6 +294,27 @@
 		{
 			"backtrace" : 1,
 			"path" : "src/util.h",
+			"sourceGroupIndex" : 1
+		},
+		{
+			"backtrace" : 1,
+			"compileGroupIndex" : 0,
+			"path" : "src/keys/mouse.c",
+			"sourceGroupIndex" : 0
+		},
+		{
+			"backtrace" : 1,
+			"path" : "src/keys/mouse.h",
+			"sourceGroupIndex" : 1
+		},
+		{
+			"backtrace" : 1,
+			"path" : "src/keys/functions.h",
+			"sourceGroupIndex" : 1
+		},
+		{
+			"backtrace" : 1,
+			"path" : "src/dwm.h",
 			"sourceGroupIndex" : 1
 		}
 	],
