+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Oct 31 07:03 EET
+Start testing: Oct 31 10:02 EET
 ----------------------------------------------------------
-End testing: Oct 31 07:03 EET
+End testing: Oct 31 10:02 EET
