+++ src/colors.h
@@ -1,11 +1,6 @@
 #ifndef colors_H
 #define colors_H
 
-// #define wal "/home/anas/.cache/wal/colors-wal-dwm.h"
-
-// #if __has_include(wal)
-// #include wal
-// #else
 static const char norm_fg[] = "#e4ded4";
 static const char norm_bg[] = "#191b1f";
 static const char norm_border[] = "#282c34";
@@ -18,12 +13,11 @@ static const char urg_fg[] = "#e4ded4";
 static const char urg_bg[] = "#313036";
 static const char urg_border[] = "#312e36";
 
-static const char *colors[][3]      = {
-    /*               fg           bg         border                         */
-    [SchemeNorm] = { norm_fg,     norm_bg,   norm_border }, // unfocused wins
-    [SchemeSel]  = { sel_fg,      sel_bg,    sel_border },  // the focused win
-    [SchemeUrg] =  { urg_fg,      urg_bg,    urg_border },
+static const char *colors[][3] = {
+        /*               fg           bg         border                         */
+        [SchemeNorm] = {norm_fg, norm_bg, norm_border}, // unfocused wins
+        [SchemeSel]  = {sel_fg, sel_bg, sel_border},  // the focused win
+        [SchemeUrg] =  {urg_fg, urg_bg, urg_border},
 };
-// #endif
 
-#endif // !colors.h
+#endif // colors_H
