+++ src/keys/functions.h
@@ -1,0 +1,46 @@
+//
+// Created by anas on 10/31/22.
+//
+
+#ifndef FUNCTIONS_H
+#define FUNCTIONS_H
+
+#include "../dwm.h"
+
+static void spawn(const Arg *arg);
+
+static void focusmon(const Arg *arg);
+
+static void fullscreen(const Arg *arg);
+
+static void focusstack(const Arg *arg);
+
+static void tagmon(const Arg *arg);
+
+static void incnmaster(const Arg *arg);
+
+static void killclient(const Arg *arg);
+
+static void movemouse(const Arg *arg);
+
+static void togglebar(const Arg *arg);
+
+static void togglesticky(const Arg *arg);
+
+static void togglefloating(const Arg *arg);
+
+static void setmfact(const Arg *arg);
+
+static void view(const Arg *arg);
+
+static void movestack(const Arg *arg);
+
+static void setlayout(const Arg *arg);
+
+static void tag(const Arg *arg);
+
+static void zoom(const Arg *arg);
+
+static void quit(const Arg *arg);
+
+#endif //FUNCTIONS_H
