+++ src/layouts/layouts.h
@@ -10,15 +10,15 @@
 
 // Layouts array
 static const Layout layouts[] = {
-	/* symbol     arrange function */
-	{ "[]=",      tile                      },    /* first entry is default */
-	{ "><>",      NULL                      },    /* no layout function means floating behavior */
-	{ "[M]",      monocle                   }, // 2
-	{ "HHH",      grid                      }, // 3
-        { "[@]",      spiral                    }, // 4
-        { "[\\]",     dwindle                   }, // 5
-        { "|M|",      centeredmaster            }, // 6
- 	{ ">M>",      centeredfloatingmaster    }, // 7
+        /* symbol     arrange function */
+        {"[]=",  tile},    /* first entry is default */
+        {"><>", NULL},    /* no layout function means floating behavior */
+        {"[M]",  monocle}, // 2
+        {"HHH",  grid}, // 3
+        {"[@]",  spiral}, // 4
+        {"[\\]", dwindle}, // 5
+        {"|M|",  centeredmaster}, // 6
+        {">M>",  centeredfloatingmaster}, // 7
 };
 
-#endif // !layouts.h
+#endif // layouts_H
