+++ src/movestack.c
@@ -1,4 +1,4 @@
-void
+static void
 movestack(const Arg *arg) {
 	Client *c = NULL, *p = NULL, *pc = NULL, *i;
